requirements.txt
mysql-connector-python
PyQt5
requests
numpy
sentence-transformers
torch
